Moving to using json from game_constants
