+++ kits/ts/simple/lux/io.ts
@@ -1,0 +1,10 @@
+/** all constants related to any input from match engine */
+export const INPUT_CONSTANTS = {
+  DONE: 'D_DONE',
+  RESEARCH_POINTS: 'rp',
+  RESOURCES: 'r',
+  UNITS: 'u',
+  CITY: 'c',
+  CITY_TILES: 'ct',
+  ROADS: 'ccd'
+};
